% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_flag.R
\name{add_flag}
\alias{add_flag}
\title{Add repelled row annotation to a heat map}
\usage{
add_flag(pheatmap, kept_labels, repel_degree = 0)
}
\arguments{
\item{pheatmap}{a pheatmap object}

\item{kept_labels}{a vector of characters, the row names to be annotated}

\item{repel_degree}{numeric, the extent to repel the annotation, the higher the
number is, the farther the adjacent row names are. Default is 0.}
}
\value{
a new pheatmap object, in which the gtable is updated.
}
\description{
Add repelled row annotation to a heat map
}
\examples{
\dontrun{

(expr = createExprData(100, 6))
library(pheatmap)
pm = pheatmap(expr, cluster_row = FALSE, scale = "row")

add_flag(pm, kept_labels = c("G001", "G023", "G059"))
add_flag(pm, kept_labels = rownames(expr)[1:5], repel_degree = 0)
add_flag(pm, kept_labels = rownames(expr)[1:5], repel_degree = .5)
add_flag(pm, kept_labels = rownames(expr)[1:5], repel_degree = 1)
}
}
