% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_add.R
\name{aes_add}
\alias{aes_add}
\title{add aes elements}
\usage{
aes_add(aes, ...)
}
\arguments{
\item{aes}{an aesthetic mapping}

\item{...}{the elements to be added to aes}
}
\description{
add aes elements
}
\examples{
\dontrun{
library(ggplot2)
col = "x3"
shp = "x4"
ae = aes(x = x1, y = x2, color = .data[[col]])
print(ae)
ae2 = aes_add(ae, shape = .data[[shp]])
print(ae2)
}
}
