% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMORplots.R
\docType{package}
\name{CMORplots}
\alias{CMORplots}
\title{CMORplots: CMOR standard themes and colour palettes for \code{ggplot}}
\description{
The CMORplots package provides \code{ggplot} themes and colour schemes for
consistent figure formatting for CMOR outputs.
}
