% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Col2Hex.R
\name{Col2Hex}
\alias{Col2Hex}
\title{Colour to hexadecimal conversion}
\usage{
Col2Hex(col, alpha = FALSE)
}
\arguments{
\item{col}{vector of any of the three kinds of \R color specifications,
    i.e., either a color name (as listed by \code{\link[grDevices]{colors}()}), a
    hexadecimal string (see Details),
    or a positive integer \code{i} meaning \code{\link[grDevices]{palette}()[i]}.}

\item{alpha}{logical value indicating whether the alpha channel (opacity)
    values should be returned.}
}
\description{
Convert R colour to hexadecimal representation.
}
\examples{
Col2Hex(1:3)
Col2Hex(c("peachpuff", "blue"), TRUE)
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
