% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExprData.R
\name{createExprData}
\alias{createExprData}
\title{Generate an example expression data}
\usage{
createExprData(nGene = 100, nSample = 15, min = 5, max = 10)
}
\arguments{
\item{nGene}{integer, number of genes}

\item{nSample}{integer, number of genes}

\item{min}{numeric, minimum value}

\item{max}{numeric, maximum value}
}
\description{
Generate an example expression data
}
\examples{
\dontrun{
createExprData(nGene = 100, nSample = 15, min = 5, max = 10)
createExprData(nGene = 200, nSample = 30, min = 5, max = 10)
}
}
