% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_histogram.R
\name{custom_histogram}
\alias{custom_histogram}
\title{Custom Histogram}
\usage{
custom_histogram(df, column, bins = 30, fill_color = "grey")
}
\arguments{
\item{df}{Data frame containing a single numeric column.}

\item{column}{Column name as a string for the data.}

\item{bins}{Integer specifying the number of bins.}

\item{fill_color}{String specifying the fill color of the histogram.}
}
\value{
A ggplot object.
}
\description{
This function generates a custom histogram, allowing users to specify the fill color and the number of bins.
}
\examples{
df <- data.frame(value = rnorm(100))
custom_histogram(df, 'value', 20, 'blue')
}
