% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_text.R
\name{dotplot_text}
\alias{dotplot_text}
\title{plot pathway/GO enrichment results with the texts in the ploting area}
\usage{
dotplot_text(
  data,
  x = "GeneRatio",
  y = "Description",
  color = "p.adjust",
  size = "Count",
  fontsize_max = 34,
  fontsize_scale = 1
)
}
\arguments{
\item{data}{a data frame from an enrichResult object}

\item{x}{character, the variable on the x axis}

\item{y}{character, the variable on the y axis}

\item{color}{character, the variable for the colors}

\item{size}{character, the variable for the point sizes}

\item{fontsize_max}{numeric, maximum fontsize (not a real fontsize, this is a relative number)}

\item{fontsize_scale}{numeric, scale to adjust fontsize}
}
\description{
plot pathway/GO enrichment results with the texts in the ploting area
}
\examples{
\dontrun{
# pathways is an enrichResult object that either clusterProfiler::enrichGO or
ReactomePA::enrichPathway function returns.
pathways = data.frame()
g = dotplot(pathways, showCategory=30)
g
dotplot_text(pathways, fontsize_max = 30, fontsize_scale = 1)
dotplot_text(g$data, fontsize_max = 24, fontsize_scale = .7)
dotplot_text(g$data, fontsize_max = 30, fontsize_scale = 1.8)
}
}
