\name{Generic function templates}
\alias{Generic function templates}
\alias{summaryplot}
\title{Generic functions}
\usage{
  summaryplot(...)
}
\arguments{
  \item{...}{parameters passed to class methods.}
}
\description{
  Generic functions for use with other packages
}
\details{
  \code{summaryplot} plot a summary of some data
}
\author{
  Ute Hahn, \email{ute@imf.au.dk}
}

