\name{is.style}
\alias{is.style}
\title{Check whether an object is a unique style list}
\usage{
  is.style(x)
}
\arguments{
  \item{x}{any \code{R} object}
}
\value{
  \code{TRUE} if \code{x} belongs to class \code{"style"},
  otherwise \code{FALSE}.
}
\description{
  Checks if an object belongs to class \code{"style"}.
}
\seealso{
  \code{\link{style}} for a constructor of this class.
}

