% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAnimation.R
\name{makeAnimation}
\alias{makeAnimation}
\title{Make animation based on gif figures}
\usage{
makeAnimation(
  imgs,
  pattern = NULL,
  path = NULL,
  fps = 2,
  outputgif = "output.gif",
  optimize = TRUE
)
}
\arguments{
\item{imgs}{image files}

\item{pattern}{if imgs is not provided, then pattern and path are used to find
the image file names, in this case, the order of the frames in the animation
is the same as the file name order}

\item{path}{the directory of the image files}

\item{fps}{integer, frames per second}

\item{outputgif}{output gif animation file}

\item{optimize}{logical, optimize the gif animation by storing only the differences
between frames. Input images must be exactly the same size. The default is TRUE.}
}
\description{
Make animation based on gif figures
}
\examples{
\dontrun{
library(ggplot2)
library(funcTools)
x = seq(0, 2*pi, length = 200)
y = sin(x)
data = data.frame(x, y)
folder = tempdir()
files = basename(tempfile())
pngs = fileNameFactory(folder, ".png")

# Plotting figures
for(mi in 1:nrow(data)){
  g = ggplot(data[mi,], aes(x, y))+
  geom_point(color = "red")+
  geom_line(data = data)+
  coord_fixed(ratio = 2, xlim = c(0, 2*pi), ylim = c(-1, 1))
  ggsave(pngs("png2/", files, "_", mi), g, width = 4, height = 4)
}

# Making the animation
gif = makeAnimation(path = paste0(folder, "/png2"), fps = 50,
              outputgif = paste0(folder, "/output.gif"))
gif

# Making the animation without compressing the images
makeAnimation(path = paste0(folder, "/png2"), fps = 50,
              outputgif = paste0(folder, "/output2.gif"), optimize = FALSE)
}
}
