% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot2D}
\alias{plot2D}
\title{Plot metadata information on PCA plot}
\usage{
plot2D(
  x,
  meta,
  colorCol = NULL,
  shapeCol = NULL,
  sizeCol = NULL,
  showSide = TRUE,
  xlab = "X",
  ylab = "Y"
)
}
\arguments{
\item{x}{n x 2 matrix, the each column indicates one dimension, the rownames(x)
must be identical to rownames(meta)}

\item{meta}{data.frame, the metadata of the samples that is going to show}

\item{colorCol}{character, which column of the meta is indicated by the color of the points}

\item{shapeCol}{character, which column of the meta is indicated by the shape of the points}

\item{sizeCol}{character, which column of the meta is indicated by the size of the points}

\item{showSide}{logical, whether to plot the distribution of the points projecting
to the x and y axis on the top and right side}

\item{xlab}{label for Y axis}
}
\description{
Plot metadata information on PCA plot
}
\examples{
\dontrun{
pca = prcomp(mtcars)
set.seed(1)
meta = data.frame(group = sample(c("A", "B"), nrow(mtcars), replace = T),
                  gear = mtcars$gear,
                  vs = mtcars$vs,
                  row.names = rownames(mtcars))
data = pca$x[,1:2]
plot2D(data, meta)
plot2D(data, meta, colorCol= "group")
plot2D(data, meta, colorCol= "group", showSide = FALSE)
plot2D(data, meta, colorCol= "gear", shapeCol = "group")
plot2D(data, meta, colorCol= "gear", shapeCol = "group", sizeCol = "vs")
}
}
