% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cmor_colours.R
\name{plot_cmor_colours}
\alias{plot_cmor_colours}
\alias{plot_cmor_colors}
\alias{list_cmor_colours}
\alias{list_cmor_colors}
\title{Show available CMOR colour palettes}
\usage{
plot_cmor_colours()

plot_cmor_colors()

list_cmor_colours()

list_cmor_colors()
}
\description{
List or plot the currently available CMOR colour palettes.
}
