% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotManhattan.R
\name{plotManhattan}
\alias{plotManhattan}
\title{Make a manhattan plot for GWAS result or DNA methylation result}
\usage{
plotManhattan(
  x,
  chr = "CHR",
  bp = "BP",
  p = "P",
  snp = "SNP",
  color = c("gray10", "gray60"),
  suggestiveline = -log10(1e-05),
  genomewideline = -log10(5e-08),
  highlight = NULL
)
}
\arguments{
\item{x}{data frame containing following information: CHR (chromosome), BP
(position of the SNP in the chromosome), P (p value), SNP (SNP name)}

\item{chr}{A string denoting the column name for the chromosome
Defaults to PLINK's "CHR"}

\item{bp}{A string denoting the column name for the chromosomal position
Defaults to PLINK's "BP"}

\item{p}{A string denoting the column name for the p-value. Defaults to PLINK's "P"}

\item{snp}{A string denoting the column name for the SNP name (rs number)
Defaults to PLINK's "SNP"}

\item{color}{A character vector indicating which colors to alternate}

\item{suggestiveline}{Where to draw a "suggestive" line. Default -log10(1e-5)
Set to FALSE to disable}

\item{genomewideline}{Where to draw a "genome-wide sigificant" line
Default -log10(5e-8). Set to FALSE to disable}

\item{highlight}{A character vector of SNPs in your dataset to highlight
These SNPs should all be in your dataset}
}
\value{
ggplot object
}
\description{
Make a manhattan plot for GWAS result or DNA methylation result
}
\examples{
\dontrun{
res = data.frame(SNP = paste0("rs", 1:15000),
                 CHR = c(rep(1, 1500), rep_len(c(2:22, "X", "Y"), 13500)),
                 P = sample(c(runif(20, 10^-9, 10^-7), runif(15000-10, 10^-7, 1)), 15000))
res$BP = unlist(tapply(res$SNP, res$CHR, function(x) 1:length(x)))
highlight = res[res$P<5e-08, "SNP"]

plotManhattan(res)
plotManhattan(res, suggestiveline = FALSE, genomewideline = FALSE)
plotManhattan(res, chr = "CHR", bp = "BP", p = "P", snp = "SNP",
              color = NULL,
              suggestiveline = -log10(1e-05),
              genomewideline = -log10(5e-08),
              highlight = highlight)
}
}
