% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetaOn2D.R
\name{plotMetaOn2D}
\alias{plotMetaOn2D}
\title{Project the samples with high dimensional features onto 2D space}
\usage{
plotMetaOn2D(
  expr,
  meta,
  type = c("PCA", "UMAP", "tSNE", "MDS", "isoMDS"),
  colorCol = NULL,
  shapeCol = NULL,
  sizeCol = NULL,
  PCs = c(1, 2),
  showSide = TRUE
)
}
\arguments{
\item{expr}{the feature matrix or data frame, where each ROW is a FEATURE, such as a gene, and
each COLUMN is a SAMPLE}

\item{meta}{data.frame, the metadata of the samples that is going to show}

\item{type}{character, one of the following values:
\code{c("PCA", "UMAP", "tSNE", "MDS", "isoMDS")}, isoMDS is a nonmetric MDS
performed using MASS package. The default is "PCA"}

\item{colorCol}{character, which column of the meta is indicated by the color of the points}

\item{shapeCol}{character, which column of the meta is indicated by the shape of the points}

\item{sizeCol}{character, which column of the meta is indicated by the size of the points}

\item{PCs}{vector of two integers, indicating which PCs are to be plotted.
This is only valid for PCA plot. The default is to plot \code{c(1, 2)} (PC1 and PC2).}

\item{showSide}{logical, whether to plot the distribution of the points projecting
to the x and y axis on the top and right side}
}
\description{
Project the samples with high dimensional features onto 2D space
}
\examples{
\dontrun{
set.seed(1)
head(mtcars)
meta = data.frame(group = sample(c("A", "B"), nrow(mtcars), replace = T),
                  gear = mtcars$gear,
                  vs = mtcars$vs,
                  row.names = rownames(mtcars))
head(meta)

# PCA
plotMetaOn2D(t(mtcars), meta, type = "PCA",
             colorCol= "gear", shapeCol = "group", sizeCol = "vs",
             PCs = c(1,2), showSide = TRUE)
# PCA (PC1 vs PC3)
plotMetaOn2D(t(mtcars), meta, type = "PCA",
             colorCol= "gear", shapeCol = "group", sizeCol = "vs",
             PCs = c(1,3), showSide = TRUE)

# UMAP
plotMetaOn2D(t(mtcars), meta, type = "UMAP",
             colorCol= "gear", shapeCol = "group", sizeCol = "vs",
             showSide = TRUE)

# tSNE
plotMetaOn2D(t(mtcars), meta, type = "tSNE",
             colorCol= "gear", shapeCol = "group", sizeCol = "vs",
             showSide = TRUE)

# MDS
plotMetaOn2D(t(mtcars), meta, type = "MDS",
             colorCol= "gear", shapeCol = "group", sizeCol = "vs",
             showSide = TRUE)

# isoMDS
plotMetaOn2D(t(mtcars), meta, type = "isoMDS",
             colorCol= "gear", shapeCol = "group", sizeCol = "vs",
             showSide = TRUE)
}
}
