% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMetaOnPCA.R
\name{plotMetaOnPCA}
\alias{plotMetaOnPCA}
\title{Plot metadata information on PCA plot}
\usage{
plotMetaOnPCA(
  pca,
  meta,
  colorCol = NULL,
  shapeCol = NULL,
  sizeCol = NULL,
  PCs = c(1, 2),
  showSide = TRUE
)
}
\arguments{
\item{pca}{the results that prcomp function returns}

\item{meta}{data.frame, the metadata of the samples that is going to show}

\item{colorCol}{character, which column of the meta is indicated by the color of the points}

\item{shapeCol}{character, which column of the meta is indicated by the shape of the points}

\item{sizeCol}{character, which column of the meta is indicated by the size of the points}

\item{PCs}{vector of two integers, indicating which PCs are to be plotted}

\item{showSide}{logical, whether to plot the distribution of the points projecting
to the x and y axis on the top and right side}
}
\description{
Plot metadata information on PCA plot
}
\examples{
\dontrun{
pca = prcomp(mtcars)
set.seed(1)
meta = data.frame(group = sample(c("A", "B"), nrow(mtcars), replace = T),
                  gear = mtcars$gear,
                  vs = mtcars$vs,
                  row.names = rownames(mtcars))
plotMetaOnPCA(pca, meta)
plotMetaOnPCA(pca, meta, colorCol= "group")
plotMetaOnPCA(pca, meta, colorCol= "group", showSide = FALSE)
plotMetaOnPCA(pca, meta, colorCol= "gear", shapeCol = "group")
plotMetaOnPCA(pca, meta, colorCol= "gear", shapeCol = "group", sizeCol = "vs")
}
}
