\name{plotTools-package}
\alias{plotTools-package}
\alias{plotTools}
\docType{package}
\title{
\packageTitle{plotTools}
}
\description{
\packageDescription{plotTools}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{plotTools}
\packageIndices{plotTools}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{plotTools}

Maintainer: \packageMaintainer{plotTools}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
