% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottools.R
\docType{package}
\name{plottools}
\alias{plottools-package}
\alias{plottools}
\title{plotools: CMOR extensions for \code{ggplot2}}
\description{
The plottools package provides extensions to \code{ggplot2} for
consistent figure formatting for CMOR outputs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rosswilson-nz/plottools}
  \item Report bugs at \url{https://github.com/rosswilson-nz/plottools/issues}
}

}
\author{
\strong{Maintainer}: Ross Wilson \email{ross.wilson@otago.ac.nz}

}
