% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Volcano plot}
\usage{
plotVolcano(
  data,
  log2fcColumn = "log2FC",
  pvalColumn = "p",
  averageColumn = "Ave",
  sampleColumn = NULL,
  log2fc = 0,
  ave = NULL,
  p = 0.05,
  label = c("sig", "none", "all", "up", "down")
)
}
\arguments{
\item{data}{either data frame or DataFrame or tibble}

\item{log2fcColumn}{character, which column (name) is log2 fold change}

\item{pvalColumn}{character, which column (name) is p value}

\item{averageColumn}{character, which column (name) is average expression}

\item{log2fc}{numeric, log2 fold change cutoff, default is 0.}

\item{ave}{numeric, average expression cutoff, default is NULL.}

\item{p}{numeric, average p value cutoff, default is 0.05}

\item{label}{one of c("sig", "none", "all", "up", "down"), whether to annotate
the genes (points), and which type of genes to be annotated. "sig": all the genes
that pass all of the cutoffs; "none": no genes; "all": all of the genes in the
plot; "up": up-regulated genes; "down": down-regulated genes.}
}
\value{
a ggplot object of a volcano plot.
}
\description{
Volcano plot
}
\examples{
\dontrun{
library(S4Vectors)

 ## DataFrame (DFrame)
data = DataFrame(gene = letters,
log2FC = runif(26, -4, 4),
p = seq(0.001, 0.8, length = 26),
Ave = rnorm(26, mean = 100, sd = 10))
plotVolcano(data)
plotVolcano(data, label = "all")
plotVolcano(data, label = "up")
plotVolcano(data, log2fc = 1, label = "up")

## data frame
data2 = as.data.frame(data)
plotVolcano(data2)

## tibble
data3 = as_tibble(data)
g = plotVolcano(data3)
g

# Add Ave as the point size
g$mapping = aes_add(aes = g$mapping, size = Ave)
g

## data.table
data4 = as.data.table(data)
plotVolcano(data4)
}
}
