% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_cmor.R
\name{scale_colour_cmor}
\alias{scale_colour_cmor}
\alias{scale_color_cmor}
\alias{scale_fill_cmor}
\title{CMOR standard colour scales}
\usage{
scale_colour_cmor(..., direction = 1, grey = FALSE, aesthetics = "colour")

scale_color_cmor(..., direction = 1, grey = FALSE, aesthetics = "colour")

scale_fill_cmor(..., direction = 1, grey = FALSE, aesthetics = "fill")
}
\arguments{
\item{...}{Passed on to \code{\link[ggplot2]{discrete_scale}} to control
name, limits, breaks, etc.}

\item{direction}{Sets the order of colours in the scale (1 for standard, -1
for reversed).}

\item{grey}{If \code{TRUE}, grey will always be used as the last group colour
(regardless of the number of groups in the data).}

\item{aesthetics}{Which aesthetic(s) should the scale work with. This can be
useful, for example, to apply the same settings to both the
\code{colour} and \code{fill} aesthetics simultaneously.}
}
\description{
Several standard colour schemes for CMOR outputs. Only qualitative
(including paired) scales are currently implemented (sequential,
diverging, and continuous scales are planned).
}
\details{
All included palettes are colour-blind friendly.

Sources:

\code{CUD}: 'Color Universal Design' palette
(https://jfly.uni-koeln.de/color/#pallet)

\code{Bright}, \code{High-Contrast}, \code{Vibrant}, \code{Muted},
\code{Medium-Contrast}, \code{Light}:
https://personal.sron.nl/~pault/#sec:qualitative

\code{Paired}: ColorBrewer 4-class PuOr
(https://colorbrewer2.org/#type=diverging&scheme=PuOr&n=4)
}
