% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot.R
\name{screeplot}
\alias{screeplot}
\title{Plot a scree plot based on the result returned by prcomp function}
\usage{
screeplot(pca)
}
\arguments{
\item{pca}{The result returned by prcomp function}
}
\description{
Plot a scree plot based on the result returned by prcomp function
}
\examples{
\dontrun{
x = matrix(rnorm(10*20), nrow= 20)
pca = prcomp(x)
screeplot(pca)
}

}
