% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrumLegend.R
\name{SpectrumLegend}
\alias{SpectrumLegend}
\alias{SizeLegend}
\title{Produce a legend for continuous gradient scales}
\usage{
SpectrumLegend(
  x0 = 0.05,
  y0 = 0.05,
  x1 = x0,
  y1 = y0 + 0.2,
  absolute = FALSE,
  legend = character(0),
  palette,
  lty = 1,
  lwd = 4,
  bty = "o",
  bg = par("bg"),
  box.lwd = par("lwd"),
  box.lty = par("lty"),
  box.col = par("fg"),
  lend = "square",
  cex = 1,
  text.col = par("col"),
  font = NULL,
  text.font = font,
  title = NULL,
  title.col = text.col[1],
  title.cex = cex[1],
  title.adj = 0.5,
  title.font = text.font[1],
  pos = 4,
  ...
)

SizeLegend(
  x0 = 0.05,
  y0 = 0.05,
  x1 = x0,
  y1 = y0 + 0.2,
  absolute = FALSE,
  legend = character(0),
  width = c(0, 1),
  scale = c("pch", "lwd"),
  lend = "square",
  bty = "o",
  bg = par("bg"),
  box.lwd = par("lwd"),
  box.lty = par("lty"),
  box.col = par("fg"),
  cex = 1,
  col = par("col"),
  text.col = par("col"),
  font = NULL,
  text.font = font,
  title = NULL,
  title.col = text.col[1],
  title.cex = cex[1],
  title.adj = 0.5,
  title.font = text.font[1],
  pos = 4,
  ...
)
}
\arguments{
\item{x0, y0, x1, y1}{Coordinates of the bottom-left and top-right end of the
bar.}

\item{absolute}{Logical specifying whether \code{x} and \code{y} values denote
coordinates (\code{TRUE}) or relative position, where (0, 0) denotes the
bottom-left of the plot area and (1, 1) the top right.}

\item{legend}{Character vector with which to label points on \code{palette}.}

\item{palette}{Colour palette to depict.}

\item{lwd, lty, lend}{Additional parameters to \code{\link[=segments]{segments()}},
controlling line style.}

\item{bty}{Character specifying the type of box to be drawn around the
legend. The allowed values are \code{"o"} (the default) and \code{"n"}.}

\item{bg}{The background colour for the legend box. (Note that this is used only if \code{bty != "n"}.)}

\item{box.lty, box.lwd, box.col}{The line type, width and colour for the legend box (if \code{bty = "o"}).}

\item{cex}{Character expansion factor relative to current \code{par("cex")}.}

\item{text.col}{Colour used for the legend text.}

\item{font, text.font}{Font used for the legend text; see \code{\link[=text]{text()}}.}

\item{title}{Text to display}

\item{title.col}{Colour for title; defaults to \code{text.col[1]}.}

\item{title.cex}{Expansion factor(s) for the title, defaults to \code{cex[1]}.}

\item{title.adj}{Horizontal adjustment for title: see the help for
\code{par("adj")}.}

\item{title.font}{Font used for the legend title.}

\item{pos, \dots}{Additional parameters to \code{\link[=text]{text()}}.}

\item{width}{Vector of length two specifying width of legend bar at base
and top.}

\item{scale}{Character string specifying whether \code{width = 1} corresponds to:
\code{"pch"}, the size of a plotting symbol with \code{pch = 1};
\code{"lwd"}, the width of a line with \code{lwd = 1}.}

\item{col}{Colour used for the width bar.}
}
\value{
A list, returned invisibly, with components:
\itemize{
\item \code{rect} A list with components:
\itemize{
\item \code{w}, \code{h}: positive numbers giving \emph{w}idth and \emph{h}eight of the
legend's box.
\item \code{left}, \code{top}: x and y coordinates of the upper left corner of the
box.
}
\item \code{text}: A list with components \code{x}, \code{y}, numeric vectors of length
\code{length(legend)}, giving the x and y coordinates of the legend's text(s).
}
}
\description{
Prints an annotated vertical bar coloured according to a continuous palette.
}
\details{
This convenience function is not yet very customizable; do file a GitHub
issue if you would value additional functionality.
}
\examples{
plot(0:1, 0:1, type = "n", frame.plot = FALSE,
     xlab = "x", ylab = "y")
SpectrumLegend(legend = c("Dark", "Middle", "Bright"),
               palette = hcl.colors(32L), lwd = 5,
               title = "Brightness")
SpectrumLegend(0.4, 0.95, 0.9, 0.95, abs = TRUE,
               legend = seq(1, 9, by = 2), palette = 1:8, pos = 1)
SizeLegend(0.8, title = "Width", legend = c("min", "max"))
}
\author{
\href{https://orcid.org/0000-0001-5660-1727}{Martin R. Smith}
(\href{mailto:martin.smith@durham.ac.uk}{martin.smith@durham.ac.uk})
}
