% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cmor.R
\name{theme_cmor_classic}
\alias{theme_cmor_classic}
\title{Classic \code{ggplot} Theme for CMOR Outputs}
\usage{
theme_cmor_classic(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{Base font size, in pts}

\item{base_family}{Base font family}
}
\description{
This is a preliminary draft of a standard CMOR ggplot theme (just directly
calls \code{theme_classic()} for now).
}
