% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_ANOVA_heatmap.R
\name{PCA_ANOVA}
\alias{PCA_ANOVA}
\title{Association between the metadata with the principle components of expression data}
\usage{
PCA_ANOVA(
  expr,
  pd,
  Npcs = NULL,
  factorTestMethod = c("anova", "kruskal"),
  scale. = TRUE
)
}
\arguments{
\item{expr}{Expression-like data (genes in rows, samples in columns)}

\item{pd}{phenodata, (also called metadata)}

\item{Npcs}{Number of principle components}

\item{factorTestMethod}{which method is used for testing factor variable}

\item{scale.}{whether to scale the expr for PCA analysis}
}
\description{
Perform statistical test (either ANOVA or kruskal test for categorical data, and a
linear model for numerical data) to associate the metadata with the principle
components of expression data. A matrix of p-values is returned, and the percentage of
variance contributed by each principle component are returned.
}
\examples{
\dontrun{
p = PCA_ANOVA(expr, pd, Npcs = 10, factorTestMethod = "anova", scale. = TRUE)
print(p)
attr(p, "percent_variance")
}

}
