% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA_ANOVA_heatmap.R
\name{plotPCA_ANOVA_Pvalue}
\alias{plotPCA_ANOVA_Pvalue}
\title{Plot a heat map for the p values from PCA_ANOVA function}
\usage{
plotPCA_ANOVA_Pvalue(
  pvalues,
  threshold = 0.05,
  filename = NULL,
  width = 6,
  height = 6,
  units = "in"
)
}
\arguments{
\item{pvalues}{a matrix of p values from PCA_ANOVA function}

\item{threshold}{threshold for p value, above this number will be set to 1,  default 0.05,}

\item{filename}{file name, default NULL,}

\item{width}{default 6,}

\item{height}{default 6,}

\item{units}{default "in"}
}
\description{
Plot a heat map for the p values from PCA_ANOVA function
}
\examples{
\dontrun{
# expr: expression data (each row is a gene, each column is a sample)
# pd: phenodata (metadata), each row is a sample, each column is a clinical parameter

expr = matrix(c(rnorm(100, 1, 1), rnorm(100, 2, 1)), ncol = 10,
dimnames = list(paste0("g", 1:20), paste0("s", 1:10)))
pd = data.frame(group = rep(c("A", "B"), each = 5),
                batch = paste0("b", c(1,2,3,1,2,3,1,2,3,1)),
                row.names = colnames(expr))

p = PCA_ANOVA(expr, pd, Npcs = 10, factorTestMethod = "anova", scale. = TRUE)
plotPCA_ANOVA_Pvalue(p)
}

}
