% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTools-package.R
\docType{package}
\name{PlotTools-package}
\alias{PlotTools}
\alias{PlotTools-package}
\title{PlotTools}
\description{
"PlotTools" is an R package that allows the legends for continuous variables
to be added to plots using the familiar functions of the default 'graphics'
package.  It also includes utility functions to manipulate irregular polygons
and locate their centres.
}
\details{
Full documentation is available \href{https://ms609.github.io/PlotTools/}{online}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ms609.github.io/PlotTools/}
  \item \url{https://github.com/ms609/PlotTools/}
  \item Report bugs at \url{https://github.com/ms609/PlotTools/issues/}
}

}
\author{
\strong{Maintainer}: Martin R. Smith \email{martin.smith@durham.ac.uk} (\href{https://orcid.org/0000-0001-5660-1727}{ORCID}) [copyright holder]

}
\keyword{internal}
